Arquivos da aplicação final.

+++ app/controllers/backoffice/admins_controller.rb
@@ -1,18 +1,25 @@
 class Backoffice::AdminsController < BackofficeController
   before_action :set_admin, only: [:edit, :update, :destroy]  
+  after_action :verify_authorized, only: [:new, :destroy]
+  after_action :verify_policy_scoped, only: [:index]
   
   def index
-    @admins = Admin.all
+    #@admins = Admin.all
+    #@admins = Admin.with_full_access
+    #@admins = Admin.with_restricted_access
+    @admins = policy_scope(Admin)   
+
   end
 
   def new
     @admin = Admin.new
+    authorize @admin
   end
 
   def create
     @admin = Admin.new(params_admin)
     if @admin.save    
-      redirect_to backoffice_admins_path, notice: "O administrador (#{@admin.email}) foi salvo com sucesso!"
+      redirect_to backoffice_admins_path, notice: I18n.t('messages.created_with', item: @admin.name)
     else
       render :new, notice: "Erro ao salvar"
     end
@@ -22,25 +29,20 @@ class Backoffice::AdminsController < BackofficeController
   end
 
   def update  
-    pwd = params[:admin][:password]
-    pwd_confirmation = params[:admin][:password_confirmation]
-
-    if pwd.blank? && pwd_confirmation.blank?
-      params[:admin].delete(:password)
-      params[:admin].delete(:password_confirmation)
-    end
-
-    if @admin.update(params_admin)   
-      redirect_to backoffice_admins_path, notice: "O administrador (#{@admin.email}) foi atualizado com sucesso!"
+   
+    if @admin.update(params_admin) 
+      AdminMailer.update_email(current_admin, @admin).deliver_now  
+      redirect_to backoffice_admins_path, notice: I18n.t('messages.updated_with', item: @admin.name)
     else
       render :edit, notice: "Erro ao atualizar"
     end
   end
 
   def destroy
-    admin_email = @admin.email
+    authorize @admin
+    admin_name = @admin.name
     if @admin.destroy
-      redirect_to backoffice_admins_path, notice: "O administrador (#{admin_email}) foi excluído com sucesso!"
+      redirect_to backoffice_admins_path, notice: I18n.t('messages.destroyed_with', item: admin_name)
     else
       render :index
     end
@@ -49,14 +51,25 @@ class Backoffice::AdminsController < BackofficeController
   
   private
 
-  def set_admin
-    @admin = Admin.find(params[:id])
-  end
+    def set_admin
+      @admin = Admin.find(params[:id])
+    end
+    
+    def params_admin  
+      if password_blank?
+        params[:admin].except!(:password, :password_confirmation)     
+      end
+      if @admin.blank?
+        params.require(:admin).permit([:name, :email, :role, :password, :password_confirmation])
+      else
+        params.require(:admin).permit(policy(@admin).permitted_attributes)
+      end
+              
+    end
+
+    def password_blank?
+      params[:admin][:password].blank? &&
+      params[:admin][:password_confirmation].blank?   
+    end
   
-  def params_admin 
-    params.require(:admin).permit(:name, :email, :password, :password_confirmation)        
-  end
-
-
- 
 end
